api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "OneRun API"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-openai>=0.3.32",
    "nanoid>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "temporalio>=1.16.0",
    "uvicorn>=0.35.0",
]
